refactor from common/OboeDebug.h to oboe/OboeDebug.h
