+++ apps/OboeTester/app/src/main/cpp/TestColdStartLatency.cpp
@@ -17,7 +17,7 @@
 #include <stdlib.h>
 #include <aaudio/AAudioExtensions.h>
 
-#include "common/OboeDebug.h"
+#include "oboe/OboeDebug.h"
 #include "common/AudioClock.h"
 #include "TestColdStartLatency.h"
 #include "OboeTools.h"
