+++ apps/OboeTester/app/src/main/cpp/TestErrorCallback.cpp
@@ -16,7 +16,7 @@
 
 #include <stdlib.h>
 
-#include "common/OboeDebug.h"
+#include "oboe/OboeDebug.h"
 #include "TestErrorCallback.h"
 
 using namespace oboe;
