+++ src/aaudio/AudioStreamAAudio.cpp
@@ -21,7 +21,7 @@
 #include "aaudio/AAudioLoader.h"
 #include "aaudio/AudioStreamAAudio.h"
 #include "common/AudioClock.h"
-#include "common/OboeDebug.h"
+#include "oboe/OboeDebug.h"
 #include "oboe/Utilities.h"
 #include "AAudioExtensions.h"
