+++ src/opensles/EngineOpenSLES.cpp
@@ -15,7 +15,7 @@
  */
 
 #include <dlfcn.h>
-#include "common/OboeDebug.h"
+#include "oboe/OboeDebug.h"
 #include "EngineOpenSLES.h"
 #include "OpenSLESUtilities.h"
